refactor(desktop): top-anchor channel intro via shared ChannelIntroBlock
Extract the empty-channel intro into one component used by both the
virtualized leading row and the non-virtualized empty state, and anchor
it at the top in both. The blank Welcome channel and the first-message
state now share the same layout, so the Welcome opener's entrance plays
with zero layout shift.
