+++ desktop/src/features/messages/ui/MessageTimeline.tsx
@@ -1,5 +1,4 @@
 import * as React from "react";
-import { Hash } from "lucide-react";
 
 import {
   isDeferredTimelineSnapshotStale,
@@ -18,6 +17,7 @@ import { channelChrome } from "@/shared/layout/chromeLayout";
 import { Spinner } from "@/shared/ui/spinner";
 import { TooltipProvider } from "@/shared/ui/tooltip";
 import { UnreadPill, unreadCountLabel } from "@/shared/ui/UnreadPill";
+import { ChannelIntroBlock, type ChannelIntro } from "./ChannelIntroBlock";
 import { TimelineSkeleton, useTimelineSkeletonRows } from "./TimelineSkeleton";
 import { TimelineMessageList } from "./TimelineMessageList";
 import type { TimelineVirtualizerApi } from "./TimelineMessageList";
@@ -111,22 +111,6 @@ type MessageTimelineProps = {
   unreadCount?: number;
   /** Per-thread unread counts keyed by thread root id. */
   threadUnreadCounts?: ReadonlyMap<string, number>;
-};
-
-type ChannelIntroAction = {
-  description?: string;
-  icon: React.ReactNode;
-  label: string;
-  onClick: () => void;
-  testId?: string;
-};
-
-type ChannelIntro = {
-  actions?: ChannelIntroAction[];
-  channelKindLabel: string;
-  channelName: string;
-  description?: string | null;
-  icon?: React.ReactNode;
 };
 
 /** Stable empty reference used as the `useDeferredValue` initial value so the
@@ -585,70 +569,7 @@ const MessageTimelineBase = React.forwardRef<
   const virtualizedLeadingContent = React.useMemo(
     () =>
       activeChannelIntro ? (
-        <div
-          className="flex w-full flex-col items-start px-3 pb-4 pt-2 text-left"
-          data-testid="message-channel-intro"
-        >
-          <div className="flex h-[60px] w-[60px] items-center justify-center rounded-2xl border border-border/70 bg-muted/40 text-muted-foreground">
-            {activeChannelIntro.icon ?? (
-              <Hash aria-hidden className="h-7 w-7" />
-            )}
-          </div>
-          <p className="mt-4 max-w-2xl truncate text-xl font-semibold leading-7 tracking-tight text-foreground">
-            #{activeChannelIntro.channelName}
-          </p>
-          <p className="mt-1 max-w-2xl text-sm leading-5 text-muted-foreground">
-            This is the beginning of the{" "}
-            <span className="font-medium text-foreground">
-              {activeChannelIntro.channelKindLabel}
-            </span>
-            .
-          </p>
-          {activeChannelIntro.description ? (
-            <p className="mt-2 max-w-xl text-sm leading-5 text-muted-foreground">
-              {activeChannelIntro.description}
-            </p>
-          ) : null}
-          {activeChannelIntro.actions?.length ? (
-            <div className="mt-4 flex max-w-full flex-nowrap gap-3 overflow-x-auto pb-1">
-              {activeChannelIntro.actions.map((action) => (
-                <button
-                  className={cn(
-                    "flex shrink-0 border border-border/70 bg-background/70 text-left transition-colors hover:bg-muted/60 focus-visible:outline-hidden focus-visible:ring-2 focus-visible:ring-ring",
-                    action.description
-                      ? "h-56 w-[13.75rem] flex-col rounded-2xl p-4"
-                      : "h-28 w-64 flex-col rounded-2xl p-4",
-                  )}
-                  data-testid={action.testId}
-                  key={action.label}
-                  onClick={action.onClick}
-                  type="button"
-                >
-                  <span
-                    className={cn(
-                      "flex shrink-0 items-center justify-center rounded-full bg-muted/70 text-muted-foreground",
-                      action.description
-                        ? "h-12 w-12 [&_svg]:h-6 [&_svg]:w-6"
-                        : "h-10 w-10 [&_svg]:h-4 [&_svg]:w-4",
-                    )}
-                  >
-                    {action.icon}
-                  </span>
-                  <span className="mt-auto min-w-0">
-                    <span className="block whitespace-normal break-words text-base font-medium leading-6 text-foreground">
-                      {action.label}
-                    </span>
-                    {action.description ? (
-                      <span className="mt-1 block whitespace-normal break-words text-sm leading-5 text-muted-foreground">
-                        {action.description}
-                      </span>
-                    ) : null}
-                  </span>
-                </button>
-              ))}
-            </div>
-          ) : null}
-        </div>
+        <ChannelIntroBlock className="pb-4 pt-2" intro={activeChannelIntro} />
       ) : activeDirectMessageIntro ? (
         <div
           className="mb-2 flex w-full flex-col items-start px-3 pb-2 pt-2 text-left"
@@ -844,96 +765,12 @@ const MessageTimelineBase = React.forwardRef<
                 ) : null}
 
                 {activeChannelIntro ? (
-                  <div
-                    className="mt-auto flex w-full flex-col items-start px-3 py-2 text-left"
-                    data-testid="message-channel-intro"
-                  >
-                    <div
-                      className="flex h-[60px] w-[60px] items-center justify-center rounded-2xl border border-border/70 bg-muted/40 text-muted-foreground"
-                      data-testid="message-channel-intro-icon"
-                    >
-                      {activeChannelIntro.icon ?? (
-                        <Hash aria-hidden className="h-7 w-7" />
-                      )}
-                    </div>
-                    <p className="mt-4 max-w-2xl truncate text-xl font-semibold leading-7 tracking-tight text-foreground">
-                      #{activeChannelIntro.channelName}
-                    </p>
-                    <p className="mt-1 max-w-2xl text-sm leading-5 text-muted-foreground">
-                      This is the beginning of the{" "}
-                      <span className="font-medium text-foreground">
-                        {activeChannelIntro.channelKindLabel}
-                      </span>
-                      .
-                    </p>
-                    {activeChannelIntro.description ? (
-                      <p className="mt-2 max-w-xl text-sm leading-5 text-muted-foreground">
-                        {activeChannelIntro.description}
-                      </p>
-                    ) : null}
-                    {activeChannelIntro.actions?.length ? (
-                      <div className="mt-4 flex max-w-full flex-nowrap gap-3 overflow-x-auto pb-1">
-                        {activeChannelIntro.actions.map((action) => {
-                          const hasDescription = Boolean(action.description);
-
-                          return (
-                            <button
-                              className={cn(
-                                "flex shrink-0 border border-border/70 bg-background/70 text-left transition-colors hover:bg-muted/60 focus-visible:outline-hidden focus-visible:ring-2 focus-visible:ring-ring",
-                                hasDescription
-                                  ? "h-56 w-[13.75rem] flex-col rounded-2xl p-4"
-                                  : "h-28 w-64 flex-col rounded-2xl p-4",
-                              )}
-                              data-testid={action.testId}
-                              key={action.label}
-                              onClick={action.onClick}
-                              type="button"
-                            >
-                              <span
-                                className={cn(
-                                  "flex shrink-0 items-center justify-center rounded-full bg-muted/70 text-muted-foreground",
-                                  hasDescription
-                                    ? "h-12 w-12 [&_svg]:h-6 [&_svg]:w-6"
-                                    : "h-10 w-10 [&_svg]:h-4 [&_svg]:w-4",
-                                )}
-                                data-testid={
-                                  action.testId
-                                    ? `${action.testId}-icon`
-                                    : undefined
-                                }
-                              >
-                                {action.icon}
-                              </span>
-                              <span className="mt-auto min-w-0">
-                                <span
-                                  className="block whitespace-normal break-words text-base font-medium leading-6 text-foreground"
-                                  data-testid={
-                                    action.testId
-                                      ? `${action.testId}-title`
-                                      : undefined
-                                  }
-                                >
-                                  {action.label}
-                                </span>
-                                {action.description ? (
-                                  <span
-                                    className="mt-1 block whitespace-normal break-words text-sm leading-5 text-muted-foreground"
-                                    data-testid={
-                                      action.testId
-                                        ? `${action.testId}-description`
-                                        : undefined
-                                    }
-                                  >
-                                    {action.description}
-                                  </span>
-                                ) : null}
-                              </span>
-                            </button>
-                          );
-                        })}
-                      </div>
-                    ) : null}
-                  </div>
+                  /* Top-anchored like the virtualized leading row, so the
+                     first message arrives below with zero layout shift. */
+                  <ChannelIntroBlock
+                    className="py-2"
+                    intro={activeChannelIntro}
+                  />
                 ) : null}
 
                 {showGenericEmpty ? (
